iota_light_client::proof

Function verify_proof

pub fn verify_proof(committee: &Committee, proof: &Proof) -> Result<()>

Verify a proof against a committee. A proof is valid if it certifies the checkpoint summary and optionally includes transaction evidence to certify objects and events.

If the result is Ok(()) then the proof is valid. If Err is returned then the proof is invalid and the error message will describe the reason. Once a proof is verified it can be trusted, and information in targets as well as checkpoint_summary or contents_proof can be trusted as being authentic.

The authoritative committee is required to verify the proof. The sequence of committees can be verified through a Committee proof target on the last checkpoint of each epoch, sequentially since the first epoch.