iota_light_client

Module utils

Structs

CheckpointsList

Config

RemotePackageStore

Enums

SCommands

Functions

check_and_sync_checkpoints

download_checkpoint_summary

download_checkpoint_summary_from_fullnode

download_checkpoint_summary_from_object_store

extract_verified_effects_and_events

get_full_checkpoint

get_full_checkpoint_from_client

get_full_checkpoint_from_object_store

get_verified_effects_and_events

get_verified_object

read_checkpoint

read_checkpoint_general

read_checkpoint_list

sync_checkpoint_list_to_latest

Run binary search to for each end of epoch checkpoint that is missing between the latest on the list and the latest checkpoint.

write_checkpoint

write_checkpoint_general

write_checkpoint_list