Module iota_light_client::utils

Structs

CheckpointsList

Config

RemotePackageStore

Enums

SCommands

Functions

check_and_sync_checkpoints

download_checkpoint_summary

extract_verified_effects_and_events

get_full_checkpoint

get_verified_effects_and_events

get_verified_object

read_checkpoint

read_checkpoint_general

read_checkpoint_list

read_checkpoint_list_from_config

sync_checkpoint_list_to_latest

Run binary search to for each end of epoch checkpoint that is missing between the latest on the list and the latest checkpoint.

write_checkpoint

write_checkpoint_general

write_checkpoint_list